fix(web): reconcile the settings work with settings discovery
Rebasing onto #2281 left three quality regressions, all in code the rebase
touched.

Pointing the System and Account headers at their first page dropped the last
use of their prefix constants, so those imports went unused.

The route table crossed its line budget, so the redirect stub and the
recording hook move to settings-route-helpers.tsx. The table is a table; a
component that only calls router.replace and a hook that only records a path
were the two things in it that were not routes.

The footer test's gear cases move into their own describe — with their own
cleanup, which the first split forgot — and two literals it now repeats four
times become constants.
